requirements.txt
Flask
line-bot-sdk
keras
numpy
Pillow
gunicorn
tensorflow